Split section handlers to different files
