+++ assets/dev/js/frontend/elements-handler.js
@@ -7,7 +7,7 @@ import toggleHandler from './handlers/toggle';
 import videoHandler from './handlers/video';
 import imageCarouselHandler from './handlers/image-carousel';
 import textEditorHandler from './handlers/text-editor';
-import sectionHandler from './handlers/section';
+import sectionHandler from './handlers/section/section';
 
 module.exports = function( $ ) {
 	const self = this;
